Split files into different packages
