+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,15 +1,21 @@
+package duke.command;
+
+import duke.Storage;
+import duke.UI;
+import duke.task.TaskList;
+
 import java.io.FileNotFoundException;
 
-public class DoneCommand extends Command {
+public class DeleteCommand extends Command {
 
     private int taskNumber;
 
-    public DoneCommand(int taskNumber) {
+    public DeleteCommand(int taskNumber) {
         this.taskNumber = taskNumber;
     }
 
     @Override
     public void execute(TaskList tasks, UI ui, Storage storage) throws FileNotFoundException {
-        tasks.markTaskAsDone(taskNumber, storage, ui);
+        tasks.deleteTask(taskNumber, storage, ui);
     }
 }
